requirements.txt
httpx
loguru
python-dotenv